+++ ICD.Connect.Lighting.Server/LightingProcessorClientDevice.cs
@@ -96,7 +96,7 @@ namespace ICD.Connect.Lighting.Server
 			else if (port is INetworkPort)
 				(port as INetworkPort).ApplyDeviceConfiguration(m_NetworkProperties);
 
-			m_RpcController.SetPort(port);
+			m_RpcController.SetPort(port, false);
 
 			UpdateCachedOnlineStatus();
 		}
@@ -270,7 +270,7 @@ namespace ICD.Connect.Lighting.Server
 		{
 			base.ClearSettingsFinal();
 
-			m_RpcController.SetPort(null);
+			m_RpcController.SetPort(null, false);
 			m_RoomId = 0;
 
 			m_NetworkProperties.ClearNetworkProperties();
@@ -304,6 +304,17 @@ namespace ICD.Connect.Lighting.Server
 			}
 
 			ConfigurePort(port);
+		}
+
+		/// <summary>
+		/// Override to add actions on StartSettings
+		/// This should be used to start communications with devices and perform initial actions
+		/// </summary>
+		protected override void StartSettingsFinal()
+		{
+			base.StartSettingsFinal();
+
+			m_RpcController.Start();
 		}
 
 		#endregion